+++ R/MANOVA_simCI.R
@@ -53,7 +53,7 @@ simCI <- function(object, contrast, contmat = NULL, type = NULL,
     refit.formula <- as.formula(paste(form1, "~", factor))
     object <- object$modelcall(refit.formula, data = object$input$data, iter = object$input$iter, 
                                resampling = object$input$resampling, alpha = object$input$alpha,
-                               seed = object$input$seed, subject = object$input$subject, CPU = object$CPU)
+                               seed = object$input$seed, subject = object$input$subject)
     
   }